changed the code in graphchi_engine.hpp to use 64-bit calculation for computing window size. When maxwindow is negative and gets cast to unsigned int, it becomes a huge value.

+++ src/engine/graphchi_engine.hpp
@@ -728,7 +728,7 @@ namespace graphchi {
             
             if (svertex_t().computational_edges()) {
                 // Heuristic
-                set_maxwindow(membudget_mb * 1024 * 1024 / 3 / 100);
+                set_maxwindow((size_t)membudget_mb * 1024 * 1024 / 3 / 100);
                 logstream(LOG_INFO) << "Set maxwindow:" << maxwindow << std::endl;
             }
